fix: Corrigido o tipo da variavel count que causava erro

+++ src/main/java/com/desafioitau/controller/dtos/EstatisticasResponseDTO.java
@@ -1,5 +1,5 @@
 package com.desafioitau.controller.dtos;
 
-public record EstatisticasResponseDTO(Integer count, Double sum, Double avg, Double min, Double max) {
+public record EstatisticasResponseDTO(Long count, Double sum, Double avg, Double min, Double max) {
 
 }
